Update check_style.yml (#31)
I'd like to disable the automatic code formatting checks for the interiorsim/experiments folder. Here is my reasoning.
The intent of this folder is for quick-and-dirty experiments. Quick-and-dirty experiments don't demand the same level of quality as core code. For example, I want to use this folder for my sim-to-real correlation coefficient experiments (notebooks, code, data, other documentation). This code belongs in the interiorsim repo, but it is not core code.
Quick-and-dirty experiments often require a mix of code that we write ourselves and (lightly hacked versions of) third-party code. It's better to have the option of checking in third-party code as-is without comprehensive reformatting, because then it is easy to use diff to find minor modifications. If we force comprehensive reformatting on third-party code, it becomes impossible to use diff to find minor modifications. Indeed, nearly all of the formatting errors from my SRCC experiments are from third-party code that I needed to include and lightly hack to get things working.
The experiments folder can depend on other folders, but other folders should not depend on experiments. If this rule is followed, there is no danger of quick-and-dirty experiment code leaking into core code.